+++ packages/analytics-js/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.7.9](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.7.8...@rudderstack/analytics-js@3.7.9) (2024-08-16)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-cookies` updated to version `0.3.6`
+* `@rudderstack/analytics-js-common` updated to version `3.9.0`
+* `@rudderstack/analytics-js-plugins` updated to version `3.4.9`
 ## [3.7.8](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js@3.7.7...@rudderstack/analytics-js@3.7.8) (2024-08-16)
 
 ### Dependency Updates